Fix evil-ex-echo bug in :global
With autopairing on, when pressing something like :g/[]<BS>l[] (so it
ends up as :g/[l[], which is perfectly valid), the message in the echo
area doesn't disappear. As it turns out, the relevant function
disappears from pre-command-hook, probably due to autopairing. Putting
it inside the timer code seems to fix the issue.